편집 모드 sidebar 클릭 시 카드 전환 되도록 error fixed - TODO: console.log 지우기 - [error] cardsets/:id 에서 id 전환 시 state 갱신 안되는 문제
