+++ src/containers/CardsetContainer.jsx
@@ -8,12 +8,6 @@ import SideMenuBar from '../components/SideMenuBar';
 
 import { get } from '../utils';
 
-import {
-  loadCardsetInfo,
-  loadCardsetChildren,
-  loadRootCardsets,
-} from '../actions';
-
 const Wrapper = styled.div({
   width: '100vw',
   height: '100vh',
@@ -21,18 +15,6 @@ const Wrapper = styled.div({
 });
 
 export default function CardsetContainer({ id }) {
-  const dispatch = useDispatch();
-
-  dispatch(loadCardsetInfo(id));
-  dispatch(loadCardsetChildren(id));
-  dispatch(loadRootCardsets());
-
-  useEffect(() => {
-    dispatch(loadRootCardsets());
-    dispatch(loadCardsetInfo(id));
-    dispatch(loadCardsetChildren(id));
-  }, []);
-
   const menus = useSelector(get('rootCardsets'));
   const cardsetInfo = useSelector(get('cardsetInfo'));
   const cardsetChildren = useSelector(get('cardsetChildren'));
